pyproject.toml
[build-system]
requires = ["setuptools>=80.0"]
build-backend = "setuptools.build_meta"

[project]
name = "stonk_gui"
version = "0.1.0"
description = "Download stockmarket charts, data, and heatmaps"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
keywords = ["stockmarket",]
authors = [
  {name = "A. Random Developer", email = "author@example.com"}
]
maintainers = [
  { name = "A. Great Maintainer", email = "maintainer@example.com" },
]
classifiers = [
  # How mature is this project? Common values are
  #   3 - Alpha
  #   4 - Beta
  #   5 - Production/Stable
  "Development Status :: 3 - Alpha",

  # Indicate who your project is intended for
  "Intended Audience :: Hobbyists",
  "Topic :: Stockmarket :: Technical Analysis Tools",

  # Pick your license as you wish
  "License :: Free Software Foundation :: GNU General Public License",

  # Specify the Python versions you support here. In particular, ensure
  # that you indicate you support Python 3. These classifiers are *not*
  # checked by "pip install". See instead "requires-python" key in this file.
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3 :: Only",
]
dependencies = ["pillow", "PyQt5", "python-dotenv", "scikit-learn", "selenium", "tiingo", "yfinance",]

[project.optional-dependencies]
# List additional groups of dependencies here (e.g. development
# dependencies). Users will be able to install these using the "extras"
# syntax, for example:
#   $ pip install sampleproject[dev]
dev = ["black", "coverage", "pytest",]

[project.urls]
"Homepage" = "https://github.com/pypa/stonk_gui"
"Bug Reports" = "https://github.com/pypa/stonk_gui/issues"
"Funding" = "https://donate.pypi.org"
"Say Thanks!" = "http://saythanks.io/to/example"
"Source" = "https://github.com/pypa/stonk_gui/"

[project.gui-scripts]
stonk_gui = "pkg:run_gui"

[tool.setuptools]
# If there are data files included in your packages that need to be
# installed, specify them here.
#package-data = { "sample" = ["*.dat"] }
#package-dir = {"" = "src"}

[tool.black]
line-length = 120
target-version = ['py310']

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["src"]
